+++ src/algorithms/libs/rtklib/rtklib_ppp.cc
@@ -747,7 +747,6 @@ void pppoutsolstat(rtk_t *rtk, int level, FILE *fp)
     int j;
     int week;
     int nfreq = 1;
-    char id[32];
 
     if (level <= 0 || !fp)
         {
@@ -804,11 +803,11 @@ void pppoutsolstat(rtk_t *rtk, int level, FILE *fp)
                 {
                     continue;
                 }
-            satno2id(i + 1, id);
+            auto id = satno2id(i + 1);
             for (j = 0; j < nfreq; j++)
                 {
                     fprintf(fp, "$SAT,%d,%.3f,%s,%d,%.1f,%.1f,%.4f,%.4f,%d,%.0f,%d,%d,%d,%d,%d,%d\n",
-                        week, tow, id, j + 1, ssat->azel[0] * R2D, ssat->azel[1] * R2D,
+                        week, tow, id.data(), j + 1, ssat->azel[0] * R2D, ssat->azel[1] * R2D,
                         ssat->resp[j], ssat->resc[j], ssat->vsat[j], ssat->snr[j] * 0.25,
                         ssat->fix[j], ssat->slip[j] & 3, ssat->lock[j], ssat->outc[j],
                         ssat->slipc[j], ssat->rejc[j]);