+++ src/tests/unit-tests/system-parameters/glonass_gnav_nav_message_test.cc
@@ -76,10 +76,10 @@ TEST(GlonassGnavNavigationMessageTest, String1Decoder)
     Glonass_Gnav_Ephemeris gnav_ephemeris;
 
     // Fill out ephemeris values for truth
-    gnav_ephemeris.d_P_1 = 15;
-    gnav_ephemeris.d_t_k = 7560;
+    gnav_ephemeris.d_P_1 = 0.;
+    gnav_ephemeris.d_t_k = 7560.;
     gnav_ephemeris.d_VXn = -0.490900039672852;
-    gnav_ephemeris.d_AXn = 0;
+    gnav_ephemeris.d_AXn = 0.;
     gnav_ephemeris.d_Xn = -11025.6669921875;
 
     // Call target test method